refactor(collection): finalize thin facade and update design doc
- Delegate remaining methods (list, get_recent, etc.) to managers
- Extract link replacement helpers to DocumentManager
- Cleanup imports and remove redundant code in Collection
- Update docs/design.md to reflect the new modular architecture
